extract isValidValue + applySchemaDefaultsToBlock to schemaValidation.js
Pure, dependency-free module so other tools (CI scripts, test runners,
content validators) can reuse the validation logic without pulling in
Volto's registry / React contexts / blockPath helpers.

schemaInheritance.js re-exports for backward compat. Single source of
truth: schemaValidation.js. No behavior change.
